Implement a play-by-play commentary classifier
This is to extract play-by-play commentaries from transcribed soccer
videos. The idea is to use a manually annotated transcript to bootstrap
the loop of training a classifier, using it to annotate more
transcripts, manually fix them, and use the newly annotated data to
train a better classifier until we're satisfied with the performance.